Added more properties to schema
* Added write_only and read_only to Property schema
* Added max_items and min_items to Array schema

Now unnamed struct and unnamed enums which has more than 1 field will add max_items and min_items property
to the array. The inserted value is exactly the amount of fields in the enum variant or in the unnamed struct.
This is necessary so that the serde serialization library is able to serialize and deserialize the objects correctly.