Add value_type to allow type override
Fix issue regarding thrid-party types which are not components nor they are primitive types. Since
the we cannot know what such types are by looking type type itself and those types possibly can be
presented in many forms we need to allow user to override the type in such cases.

* Add possibility to override type for named fields and unnamed field structs
* Add tests
* Update docs recardingly