Added redirects in the warp servers from `/swagger-ui` to `/swagger-ui/`
The ui only works if you go to `/swagger-ui/` with a tailing `/`.
If you browse to `/swagger-ui`, you get an unhelpful blank page,
due to the relative links for JS, CSS, et al files.
My investigation indicates that the best way to fix this in `warp`
is to manually create a redirect.
So I updated the `warp` examples to have this
(I am unsure if the other servers have the same problem)

I considered using a regex on the full path
instead of passing in both the path and tail,
but I decided passing along two strings was more efficient than
running a full regex every time