Namespacing same types (#187)
Implement namespacing for components according #186. This PR 
adds support to use multiple components with same name but in 
different module with optional casting support as seen below. 
Component name used in OpenAPI spec is the full path to the 
component where `::` is replaced with `.`. Previously mutliple 
components with same name was not possible.

```rust
components(
    crate::mod1::SomeType,
    crate::mod2::SomeType as mod3::SomeType
)
```

This work was sponsored by [Arctoris](https://www.arctoris.com/).