Add support for title (#224)
Add support for `title` property ([OpenAPI Schema title](https://swagger.io/specification/#schema-object)) for struct and enum `Component` properties.

The title is useful for enums because currently the generator for rust 
clients created a struct for each variant and name them like:

    my_enum_one_of_1
    my_enum_one_of_2
    etc

So with the title you can at least control the name of each variant structure. 
There is an open issue to actually produce an enum but it's not been worked 
on for a while OpenAPITools/openapi-generator#9497.