Fix schema serde reserialization
Currently there was an issue object component properties where not
deserialized correctly from openapi spec json. The issue was in order
the serde will try to deserialize the fields and for some reason
Property was deserialized succesfully as Object breaking the schema
definition. This PR makes the Object as first item to try to
deserialize. According to the tests now when Object is the first one the
deserializing already serialized openapi json will deserialize Object
properties correctly.

fixes #175