Chore unify Object and Property component
Unify Object and Property component into one. Reasoning to this is
OpenAPI spec does not have a separate type for `object` and `property`.
Also the types were almost the same thus unifying them is reasonable
because any `property` is actually a `object` as well.

This also mitigates possibly serde serialization and deserialization
issues where the serde is unable to correctly select deserialization
implementation. Limiting the possibly choises to one is always one round
trip less.