Add `merge` functionality for `OpenApi`
Add `merge` function to `OpenApi`. This will perform shallow comparison for
paths, schemas, responses and security schemes. This means that only
name for components and path for paths is used for comparison. Tags, security
requirements and servers will be compared against the whole object. Only
not matching elements will be merged from `OpenApi` argument to the
self.

This is useful for larger APIs allowing users to easily split the
OpenAPI definition to multiple smaller OpenAPI specs and then only
merging them in root level to a single OpenAPI specification. This is
good for encapsulation and enables users not to make all module
internals public to the root module.

Also improve generally the documentation and implement `PartialEq` and
`Eq` for openapi types for easier comparison.