Fix make untagged enum object variants required
Prior to this PR untagged enum variants were not required thus resulting
optional types when schema was used to generate types for typescript
with `typescript-axios`. This commit will add `.required(name)` to each
object variant of untagged enum. This is beneficial for better
nullability checks and results more correct API.