Refactor `Type` to `TypeTree` (#408)
Refactor custom `Type` parsing logic to `TypeTree` in parameters,
headers, responses and request bodies. The original parsing logic in
`Type` was fairly limited in capabilities and superfluous in itself.

Migrating to `TypeTree` which is is already used everywhere else comes
with benefits like single source of parsing type trees and support for
vast set of different kind of types that where not supported with the
custom type parsing. `TypeTree` allows users to use almost any Rust
types instead of those few predefined ones. The change is invisible to
the users and also `TypeTree` fixes a bug in incorrect default content
type for primitive type of arrays.