Add `ToResponse` derive implementation
Add new `ToResponse` derive attribute macro. This can be used
alternatively to generate path response. It support all the attributes
which are supported in tuple response syntax to further configure the
generated response. By default the name of the `struct` which is
decorated with `ToResponse` macro will be used as a name for the
response and as response type as well.

To define multiple contents for single response one must use enum style
definition.
```rust
 #[derive(utoipa::ToResponse)]
 enum Foo {
   Foo(#[content("type")] Foo),
   Bar(#[content("type")] Bar),
 }
```

To define `Vec` or `Option` response the new type pattern must be used.
```rust
 #[derive(utoipa::ToResponse)]
 struct MyResponse(Vec<String>);
```