Add `ToResponse` derive implementation (#416)
Add new `ToResponse` derive attribute macro. This can be used
alternatively to generate path response. It support all the attributes
which are supported in tuple response syntax to further configure the
generated response. By default the name of the `struct` which is
decorated with `ToResponse` macro will be used as a name for the
response and as response type as well. 

To define simple response component where name of the repsonse and 
schema must be same one can use named field struct or enum without 
`content` attributes.
```rust
 #[derive(utoipa::ToSchema, utoipa::ToResponse)]
 struct Person {
     name: String,
 }
```

To define multiple contents for single response one must use enum style
definition.
```rust
 #[derive(utoipa::ToResponse)]
 enum Foo {
   Foo(#[content("type")] Foo),
   Bar(#[content("type")] Bar),
 }
```

To define `Vec` or `Option` response the new type pattern must be used.
```rust
 #[derive(utoipa::ToResponse)]
 struct MyResponse(Vec<String>);
```

This PR adds docs for `ToResponse` derive macro and updates docs for _OpenApi_ 
and _path_ macro. 

Add support for inlining response reference in `utoipa::path(...responses(...)...)`.
```rust
 responses(
     (status = 200, response = inline(MyResponse))
 )
```