Enhance `ToResponse` implementation
This commit will add support for unit type responses. Structs without
fields will be threated as a responses without body.

```rust
 #[derive(ToSchema, ToResponse)]
 struct PersonSuccessResponse;
```

This commit changes `ToResponse` behaviour in named field structs and
enums without `#[content]` attribute so that the fields and variants of
those types will become inlined schema for the response. Prior to this
PR `ToSchema` ignored fields and variants of those types and created a
reference to the type having same name as response. That behaviour was
not correct and was illogical since you can use new type pattern to
create a reference response.

```rust
 #[derive(ToResponse)]
 struct Person {
     name: String,
 }
```

Add new attribute `#[to_schema]` which can be used to inline schema of
unnamed field struct or enum variant which already has `#[content]`
attribute. Without `#[to_schema]` attribute response will be created
with a reference to the field type.

```rust
 #[derive(ToResponse)]
 struct PersonSuccessResponse(#[to_schema] Vec<Person>);
```