Improve docs
Add more detailed docs for modifying generated OpenApi at runtime.
Update `IntoResponses` and `ToResponse` documentation and documentation
overall.

Update `Response` and `RequestBody` extension documentation.

This commit will remove `json_reponse_ref` method from `ResponseExt`
since it at the the beginning is wrong and is not supported by the
OpenApi spec as well.

Rename the `json_component_ref` to `json_schema_ref` to make it
consistent with rest of the crate.