Enhance generic aliases with lifetimes support
This PR will add support for aliases with lifetimes. As demonstrated now
the lifetime variable can be declared or omitted from left side of
equals (=) mark. The lifetime at right side of the equals mark will be
used as a lifetime for generated Rust type alias.

```rust
 #[aliases(Paginated1<'b> = Paginated<'b, String>, Paginated2 = Paginated<'b, Value>)]
 struct Paginated<'r, TResult> {
     pub total: usize,
     pub data: Vec<TResult>,
     pub next: Option<&'r str>,
     pub prev: Option<&'r str>,
 }
```

There is still one limitation with `'static` lifetimes. E.g if one would
define alias as `Paginated3 = Paginated<'static, Value>` this would
cause a compile error. Normally this would be allowed syntax in Rust
code.

There is a need to deprecate this alias functionality in future in favor
of adding a generic solution for Rust type aliases in general.