Refactor `ToSchema` casting as for OpenApi
This commit changes the `ToSchema` trait to be similar with `ToResponse`
trait. It also introduces lifetimes to avoid unnecessary allocations.
```rust
pub trait ToSchema<'__s> {
    fn schema() -> (&'__s str, openapi::RefOr<openapi::schema::Schema>);

    fn aliases() -> Vec<(&'__s str, openapi::schema::Schema)> {
        Vec::new()
    }
}
```

Prior to this commit the name that was used to register a schema to
OpenAPI spec was defined in `#[openapi(components(schemas(...)))]`
attribute with syntax like `schemas(path::to::Schema as
path::MySchema)`. This is a bit inconvenient at least and behaved
completely opposite to how response and request bodies behave when
defined in `#[utoipa::path]` macro.

From now the name of the schema registered in OpenAPI spec will be
defined by the `ToSchema` trait itself the same way as `ToResponse` does
it. It will return a tuple of `(name, component)` pair.

This also introduces new feature `As` what is then used to define
alternative name for the schema derived from `ToSchema` if necessary.
For example the below struct would be registered to the OpenAPI with
name `path.to.Foo` instead of just `Foo` when it is added to openapi
with `#[openapi(components(schemas(Foo)))]`.
```rust
 #[derive(ToSchema)]
 #[schema(as = path::to::Foo)]
 pub struct Foo {
     name: String,
 }
```