Improve description support on `ToSchema` fields (#490)
This commit fixes the existing implementation for field description and
deprecated status. The implementation only worked for primitive type
fields.

This PR adds correct support for following cases where the description
and deprecated status will be correctly set for vec and map fields as
well.
```rust
 struct Account {
     /// Database autogenerated id
     #[deprecated]
     id: i64,
     /// Users username
     #[deprecated]
     username: String,
     /// Role ids
     #[deprecated]
     role_ids: Vec<i32>,
     /// Users
     #[deprecated]
     users: Vec<User>,
     /// Map description
     #[deprecated]
     map: HashMap<String, String>
 }
```

Note this PR does not add support for description on `$ref` types.