Add full support for nullable field detection (#498)
This PR replaces the old old behavior with `Option` fields. The existing
implementation is naive in it's logic which is prone to result incorrect
OpenAPI clients if one is not careful with nullable fields.

 ### The old implementation

Named struct field is considered required if
* it is not `Option` type
* or it has not serde `default` attribute.

Nullable had to be defined manually for the field.

 ### New implementation

Named struct field is considered required if
  * it does not have `skip_serializing_if` property
  * and it does not have `serde_with` douple_option
  * and it does not have default value provided with serde `default`
    attribute

Also nullable is now being automatically detected from `Option` fields.

Additionally this commit add support for recognizing `double_option`
from https://github.com/jonasbb/serde_with.

Most importantly `Option` fields will be considered **required** by default
after this change. This is because by default `serde` serializes `Option` as
`null` and secondly to keep the generated classes consistent with the
API doc.