Unify component schema tokenization
Prior to this commit, there was 3 separate implementation for component
schema tokenization. This was a source of trouble and suboptimal at
least.

This PR will fix this by combining all existing implementations to
single implementation. This helps to keep code DRY and makes it easier
to maintain and add more features. Also because of this, now the same
logic will be used to create schema for ToSchema schemas, parameters and
request/response bodies. This makes parameters and request/response
body schemas in par with ToSchema schemas.