Fix untagged enum unit variant support
Prior to this commit untagged enums with unit variants caused compile
error because they where counted as unnamed field variants. However this
is incorrect behavior and they need to rendered as unit structs that
render to `null` according to serde.

This commit fixes this by allowing use of `#[serde(untagged)]` attribute
on enums with unit type fields.
```rust
 #[derive(ToSchema)]
 #[serde(untagged)]
 enum ComputeRequest {
     Aggregation(AggregationRequest),
     Breakdown,
 }
```