Wip experiemental actix_web auto-repsonses
Experiemental actix_web auto-responses from handler return type.

There are noticeable drawbacks with the automatic response resolution.
And the biggest one is question "How to override response a type that is
resolved automatically? Or even ignore one?"

Currently I have no idea on how to proceed with this one. And as the
implementation is now, is too limiting, meaning it does not cooperate
well with the tuple style responses.

Any ideas are welcome.