Fix `Option<Query<T>>` type support
Fix `Option<Query<T>>` for primitive types. Prior to this commit
defining before mentioned type as an argument to path operation handler
caused a compile error in 3.4.0 release. This bug has been there before
but it manifested only after the changes regarding automatic query
parameter detection was made.

This commit allows optional query parameters for handler function:
```rust
 async fn get_item(id: Option<Query<u32>>) {}
```

**Note!** The automatic query parameter detection does not work for
primitive types on `axum` and in `actix_web` frameworks because it has
never been implemented. They will always be expected to work with
`IntoParams` trait. This said with these frameworks the primitive query
parameters if not used with `IntoParams` must be declared manually to
the `#[utoipa::path(params(...))]` macro as input arguments.