Add crate for serving Scalar via utoipa (#892)
This is a PR for adding [Scalar](https://scalar.com/) API visualization to the utoipa ecosystem. Scalar is yet another web application for OpenAPI specifications, but I prefer it over Redoc/Rapidoc since it shows the model types too! 

Most of the things here were mimicked after the Redoc PR (#720).

- The crate has been added to the CI/publish/test scripts
- The todo-axum/todo-actix/todo-rocket examples now serve Scalar too under /scalar
- The crate has been implemented for axum, actix-web, and rocket