Add support for addtional tags via `tags`
Add support for additonal tags via `tags` attribute on `utoipa::path`
macro. All the tags will be appended to the tags will be appended to the
tags of the operation. This means that the default tag (the
handler function name by default) will be added if not overridden with
the `tag` attribute.

```rust
 #[utoipa::path(
    get,
    tag = "some_tag",
    tags = ["addtional tag 1", "another one"]
 )]
 fn handler() {}
```