Fix compile error propagation (#929)
Fix compile error propagation via comprehensive use of
`ToTokensDiagnostics` trait. `ToTokensDiagnostics` was introduced in
PR #920 to allow error propagation up in the call stack for error
reporting. However this initial implementation was too limited which
resulted invalid compile errors in case of error. This implementation
aims to fix the shortcomings of the initial implementation.

Fixes #927