Enhance OpenApi nesting with tags support
This commits enhances the `OpenApi` nesting support by adding possiblity
to define `tags [...]` for the nested OpenApi endpoints. By default the
fully qualified path to the nested `OpenApi` will become the tag for the
endpoints if provided.

This commit also makes it necessary to define the argument names within
the nest tuple to be constant across the utoipa macros and more
readable. `Path` and `api` arguments are necesary.

Supported nesting syntax:
```rust
(path = "/path/to/nest/api", api = path::to:NestableApi, tags = [...])
```

**Breaking!** This is breaking change as this changes the `utoipa::Path`
trait syntax to following. Prior this commit the the `path_item`
function took `Option<&str>`parameter to define default `tag` for the
path operation but this is no longer necessary.
```rust
trait Path {
    fn path() -> String;
    fn path_item() -> PathItem;
}
```

Follow up PR for #930

Resolves #445 Resolves #872