Add support for description and summary overriding (#948)
Add `description = ...` and `summary = ...` attributes for
`#[utoipa::path(...)]` attribute macro to allow description and summary
overriding. When these attributes are defined the values are not
resolved from the doc comment above the function.

The value can be either literal string or expression e.g. reference to a
`const` variable or `include_str!(...)` statement.

Relates to #802

Resolves #439 Resolves #781