Fix find actual request body TypeTree
Prior to this commit the find actual request body did not work for
uknown types. This PR changes the implmentation so that it does not care
about the uknown wrapper types and will try to search from it's children
if it has generics.

The example demostrates `Validated<T>` custom wrapper type being used on
body type as an out most wrapper. Before this gave a bad error but is
now supported.

```rust
 #[utoipa::path()]
 #[post("/item")]
 async fn post_item(_body: Validated<Json<Item>>) -> Json<Item> {
     Json(Item(String::new()))
 }
```