Remove old 3.0 [u8] as octet-stream behaviour
In OpenAPI 3.0 the request body and response body with type `string` and
format `binary`, `byte`, `base64` was considered a
`application/octet-stream`. However the OpenAPI 3.1 does not require
such and and the schema can be completely ignored in such cases.

This commit will remove the old behaviour where [u8] was rendered as
`type: string` `format: binary` for recognizing
`application/octet-stream`. From now on the content schema will be
resolved as is which means that [u8] will be rendered as `type: array`
of `items: int32`.