fix(utoipa-gen): use to_owned instead of to_string in expanded code
Taking something that is a string and converting it to a string
using `to_string()` kind of misses the point of why we are doing the
conversion in the first place, and also fails to document this to the
reader. `to_owned()` fully captures the reason that a conversion is
required at a particular spot: borrowed (`&str`) to owned (`String`).

Some people enable this clippy lint in order to enforce this style:

- https://rust-lang.github.io/rust-clippy/master/index.html#/str_to_string

Code generated by utoipa macro is triggering this lint.
To avoid that, this patch is replacing the `to_string()` usages in
generated code by `to_owned()`.

This patch is not changing any other `to_string()` because the utoipa
project itself does not need to adhere to this style.
The goal is just to reduce friction for downstream consumers.