Fix respect `required` attribute
Prior to this PR the `required` attribute was not correctly respected in
all cases. This was due to the fact that the `required` attribute had the
same weight with the default rule that is checked from the option and some
other serde attributes. Even this works in most cases but lacks control
over explicitly defining `required` to false when type is not
optional.

This PR fixes this by making clear separation between the default rule
and the manual `required` attribute with preference on the manually
defined `required` attribute.

Fixes #882