Fix default tag logic for paths
The default tag for paths is resolved from the module path of the handler
or `OpenApi` if the type in question is a nested `OpenApi` document. This
was supposed to be only the case when no additional tags are provided for the
paths. However it was resolved in all cases and there was no way to
"not to use" module path as a tag. This commit fixes this and the
module path will be only used if there are no other tags defined.

This commit also changes the old behavior where _`crate`_ was used as
a tag in case there was no module path available. From now on if there is
no module path there will be no default tag added to any of the paths.
This will result the paths to be rendered under _`default`_ tag in a
Swagger UI.

Fixes #978