Wip make `as` attribute authoritative name
This PR aims to change the logic how the prefix is defined for the
request bodies and response bodies throughout the library. Prior to this
the name is resolved from the value provided to the `request_body` or
response `body` attribute. However this current logic becomes cumbersome
when user needs to always provide the correct path to the attribute
value that correspond the name in the OpenAPI specification.

This changes it so that the name itself will be resolved from the name
of the schema itself thus eliminating the need for manual path
prefixing. If schema itself is provided with `as` attribute then the
name will be resolved from the same `as` attribute value.

Fixes #993