Chore change the operations implementation.
As stated here #972 the `operations` and `extensions` were both map.
This caused deserializing issues where operations where mistakenly
deserialized to extensions as well.

This commit removes the `operations` map and changes them to static
fields as they would be in the specification.
https://spec.openapis.org/oas/latest.html#fixed-fields-6

This is a breaking change and `preserver_path_order` feature changes so
that all operations under same path will always be serialized in
constant order according to fields order in the specification (link
above).

Fixes #972