Chore unify request body and ext request body
This commit unifies to logic for parsed request body and ext request
body from handler function argument. Prior to this commit they had
separate implementations for `ToTokens` but that is changed in this
commit for being overly superfluous and error prone. Now they share same
`ToTokens` logic where only difference is how the `ComponentSchema` is
resolved for the different types of request bodies.