List only utoipa related changes in its CHANGELOG
Update recent unreleased changes of `utoipa` to list only `utoipa`
related changes. Old releases are not changed. This is to remove the
duplicate entries of items that occur in `utoipa-gen` changelog as well.
Add note to indicate that there is correlation between the two crates
and one might be interested to look into changes introduced in the
`utoipa-gen` as well.