Fix impl `ToSchema` for container types
This PR fixes implementation of `ToSchema` on container types like
`Vec<T>` or `Option<T>`. Prior to this commit it was not implemented
which resulted errors when such types was used as generic argument.

This commit fixes this which correctly allows following syntax.
```rust
 #[derive(ToSchema)]
 pub struct FooStruct<B> {
     pub foo: B,
 }

 #[derive(ToSchema)]
 enum FoosEnum {
     ThingNoAliasOption(FooStruct<Option<i32>>),
     FooEnumThing(#[schema(inline)] FooStruct<Vec<i32>>),
     FooThingOptionVec(#[schema(inline)] FooStruct<Option<Vec<i32>>>),
     FooThingLinkedList(#[schema(inline)] FooStruct<std::collections::LinkedList<i32>>),
     FooThingBTreeMap(#[schema(inline)] FooStruct<std::collections::BTreeMap<String, String>>),
     FooThingHashMap(#[schema(inline)] FooStruct<std::collections::HashMap<i32, String>>),
     FooThingHashSet(#[schema(inline)] FooStruct<std::collections::HashSet<i32>>),
     FooThingBTreeSet(#[schema(inline)] FooStruct<std::collections::BTreeSet<i32>>),
 }
```

**Note!** Currently generic argument cannot be `slice`, `array` or
`tuple`. That is a hard limitation of current `TypeTree` implementation
and would need some work before such types were supported.

Relates #1034

Fixes #1099