optional Header Parameters only set required=false
When using Option<Type> in a Header parameter the current behavior for
the resulting Parameter Schema is to set required false, but instead of
just setting the Schema to Type it Wraps Type into OneOf 0: null 1:
type. This is unexpected and could be considered a bug, give that
optional Header Parameters should be just not sent if their value is
null. If this behavior is intended it should be done more explicitly.