chore: add Send+Sync to returned Boxed errors
The errors possible to return by `serve()` seem to be implementing
Send+Sync, so this just adds these auto traits to the function
signature. This is useful for passing the errors around in
web applications for instance to display it to a user eventually in HTTP
response, or to wrap it inside anyhow::Error (which requires the inner
objects to be Send+Sync).