Add `serde` to the helper attributes for `ToSchema`
This tells the compiler that the `ToSchema` macro looks at `#[serde]`
attributes. Without it, structs annotated with `#[serde]` won't compile
unless the struct also derives Serialize or Deserialize.

In cases where the Serialize/Deserialize are implemented manually but
the schema can still be derived, it may be useful to use serde
attributes that don't have a `#[schema]` equivalent, such as
`#[serde(untagged)]`.